Refactor number scrubber
